Fix for r506 commit.
This bug only affects the Windows build.  The patch from revision 506 (addition of file size to file provider API) does not work with absolute path names.  The standard file provider normalizes its UTF-8 path name.  Prior to the r506 patch the platform file class called FileSystem::getFileSize with the argument from the caller.  After the r506 patch the platform class calls the file provider to get the file size.  The standard file provider uses its normalized path name as an argument to FileSystem::getFileSize.  This was a problem if the path was absolute because the normalization added a UNC prefix, which getFileSize does not like.  The standard file provider now passes the original path value that was supplied by the caller.